+++ parachain/Cargo.toml
@@ -18,6 +18,7 @@ sp-io = { git = "https://github.com/paritytech/substrate", branch = "cumulus-bra
 lazy_static = { version = "1.4.0", optional = true }
 parking_lot = { version = "0.10.0", optional = true }
 log = { version = "0.4.8", optional = true }
+polkadot-primitives = { path = "../primitives", default-features = false }
 
 [target.'cfg(not(target_os = "unknown"))'.dependencies]
 shared_memory = { version = "0.10.0", optional = true }
@@ -44,4 +45,5 @@ std = [
 	"sp-externalities",
 	"sc-executor",
 	"sp-io",
+	"polkadot-primitives/std",
 ]
