+++ primitives/Cargo.toml
@@ -11,13 +11,13 @@ primitives = { package = "sp-core", git = "https://github.com/paritytech/substra
 inherents = { package = "sp-inherents", git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
 application-crypto = { package = "sp-application-crypto", git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
 sp-api = { git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
+sp-runtime-interface = { git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
 sp-version = { git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
 rstd = { package = "sp-std", git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
 runtime_primitives = { package = "sp-runtime", git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
-polkadot-parachain = { path = "../parachain", default-features = false }
 trie = { package = "sp-trie", git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
 bitvec = { version = "0.15.2", default-features = false, features = ["alloc"] }
-babe = { package = "pallet-babe", git = "https://github.com/paritytech/substrate", branch = "cumulus-branch", default-features = false }
+derive_more = { version = "0.99.2" }
 
 [dev-dependencies]
 sp-serializer = { git = "https://github.com/paritytech/substrate", branch = "cumulus-branch" }
@@ -31,11 +31,10 @@ std = [
 	"inherents/std",
 	"trie/std",
 	"sp-api/std",
+	"sp-runtime-interface/std",
 	"rstd/std",
 	"sp-version/std",
 	"runtime_primitives/std",
 	"serde",
-	"polkadot-parachain/std",
 	"bitvec/std",
-	"babe/std"
 ]
